Add check for requesting more vouchers than you can afford
Prints a suitable error message when you can't afford the number of
vouchers requested. Also ignores requests for 0 vouchers.

N.B.: The vouchers were never actually produced in the situation, the
Java code checks for this, but this is just to fix what it displays to
the user on the web page.